# KAFKA-10684; Avoid additional envelope copies during network transmission