--- a/clients/src/main/java/org/apache/kafka/common/protocol/ByteBufferAccessor.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/ByteBufferAccessor.java
@@ -110,7 +110,7 @@ public class ByteBufferAccessor implements Readable, Writable {
 
     @Override
     public void writeByteBuffer(ByteBuffer src) {
-        buf.put(src);
+        buf.put(src.duplicate());
     }
 
     @Override