--- a/clients/src/main/java/org/apache/kafka/common/protocol/Message.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/Message.java
@@ -48,12 +48,19 @@ public interface Message {
      *                      by this software.
      */
     default int size(ObjectSerializationCache cache, short version) {
-        MessageSize size = new MessageSize();
-        messageSize(size, cache, version);
+        MessageSizeAccumulator size = new MessageSizeAccumulator();
+        addSize(size, cache, version);
         return size.totalSize();
     }
 
-    void messageSize(MessageSize size, ObjectSerializationCache cache, short version);
+    /**
+     * Add the size of this message to an accumulator.
+     *
+     * @param size          The size accumulator to add to
+     * @param cache         The serialization size cache to populate.
+     * @param version       The version to use.
+     */
+    void addSize(MessageSizeAccumulator size, ObjectSerializationCache cache, short version);
 
     /**
      * Writes out this message to the given Writable.