--- a/clients/src/main/java/org/apache/kafka/common/protocol/RecordsWritable.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/RecordsWritable.java
@@ -115,6 +115,7 @@ public class RecordsWritable implements Writable {
         ByteUtils.writeVarlong(i, buffer);
     }
 
+    @Override
     public void writeRecords(BaseRecords records) {
         flush();
         sendConsumer.accept(records.toSend(dest));