--- a/clients/src/main/java/org/apache/kafka/common/protocol/Writable.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/Writable.java
@@ -18,6 +18,8 @@
 package org.apache.kafka.common.protocol;
 
 import org.apache.kafka.common.UUID;
+import org.apache.kafka.common.record.BaseRecords;
+import org.apache.kafka.common.record.MemoryRecords;
 
 import java.nio.ByteBuffer;
 
@@ -39,6 +41,15 @@ public interface Writable {
         short version
     ) {
         message.write(this, serializationCache, version);
+    }
+
+    default void writeRecords(BaseRecords records) {
+        if (records instanceof MemoryRecords) {
+            MemoryRecords memRecords = (MemoryRecords) records;
+            writeByteBuffer(memRecords.buffer());
+        } else {
+            throw new UnsupportedOperationException("Unsupported record type " + records.getClass());
+        }
     }
 
     default void writeUUID(UUID uuid) {