… clients/src/main/java/org/apache/kafka/common/requests/RequestContext.java
… clients/src/main/java/org/apache/kafka/common/requests/RequestHeader.java
… clients/src/main/java/org/apache/kafka/common/requests/ResponseHeader.java
… clients/src/main/java/org/apache/kafka/common/security/authenticator/SaslServerAuthenticator.java
… clients/src/main/resources/common/message/EnvelopeRequest.json
… clients/src/test/java/org/apache/kafka/common/message/RecordsSerdeTest.java
… clients/src/test/java/org/apache/kafka/common/protocol/RecordsWritableTest.java
… clients/src/test/java/org/apache/kafka/common/protocol/SendBuilderTest.java
… clients/src/test/java/org/apache/kafka/common/record/MultiRecordsSendTest.java
… clients/src/test/java/org/apache/kafka/common/requests/AbstractResponseTest.java
… clients/src/test/java/org/apache/kafka/common/requests/ByteBufferChannel.java
… clients/src/test/java/org/apache/kafka/common/requests/ByteBufferChannelTest.java
… clients/src/test/java/org/apache/kafka/common/requests/EnvelopeRequestTest.java
… clients/src/test/java/org/apache/kafka/common/requests/EnvelopeResponseTest.java
… clients/src/test/java/org/apache/kafka/common/requests/RequestContextTest.java
… clients/src/test/java/org/apache/kafka/common/requests/RequestHeaderTest.java
… clients/src/test/java/org/apache/kafka/test/TestUtils.java
… core/src/main/scala/kafka/network/RequestChannel.scala
… core/src/main/scala/kafka/network/SocketServer.scala
… core/src/main/scala/kafka/server/BrokerToControllerChannelManagerImpl.scala
… core/src/main/scala/kafka/server/ForwardingManager.scala
… core/src/main/scala/kafka/tools/TestRaftRequestHandler.scala
… core/src/test/scala/unit/kafka/server/CreateTopicsRequestWithForwardingTest.scala
… core/src/test/scala/unit/kafka/server/KafkaApisTest.scala
… generator/src/main/java/org/apache/kafka/message/MessageDataGenerator.java
… generator/src/main/java/org/apache/kafka/message/MessageGenerator.java