This patch creates a new `SendBuilder` class which allows us to avoid copying "bytes" types when transmitting an api message over the network. This is used in `EnvelopeRequest` and `EnvelopeResponse` to avoid copying the embedded data.

The patch also contains a few minor cleanups such as moving envelope parsing logic into `RequestContext`.

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)