@hachikuji 2020-11-05
@chia7712 I think it could. I was actually thinking about #9401 when I was working on it. As far as I can tell, the client currently copies the whole produce buffer when the request reaches `NetworkClient`. One remaining thing to do here I think is converge `SendBuilder` and `RecordsWritable`, which are using similar tricks.