@hachikuji 2020-11-06
@chia7712 Thanks for the reviews. I pushed an update to address your comments. I was a little concerned about the garbage created from the `MessageSize` objects, so I changed the generated code to use an accumulator pattern instead. Not sure it was really worth it, but let me know what you think.