@hachikuji 2020-11-06
Here are updated benchmarks for `FetchResponseBenchmark`.

Trunk:
```
Benchmark                                          (partitionCount)  (topicCount)  Mode  Cnt        Score       Error  Units
FetchResponseBenchmark.testSerializeFetchResponse                 3            10  avgt   15     6037.058 ±    41.647  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                 3           500  avgt   15   296913.061 ±  2548.748  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                 3          1000  avgt   15   607466.969 ±  5149.176  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10            10  avgt   15    17971.239 ±   141.934  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10           500  avgt   15   888464.188 ±  6341.169  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10          1000  avgt   15  1938108.325 ± 32056.839  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20            10  avgt   15    34217.206 ±   234.166  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20           500  avgt   15  1830593.478 ± 22601.172  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20          1000  avgt   15  4983809.392 ± 38217.505  ns/op
```

This patch:
```
Benchmark                                          (partitionCount)  (topicCount)  Mode  Cnt        Score       Error  Units
FetchResponseBenchmark.testSerializeFetchResponse                 3            10  avgt   15     5686.644 ±    92.892  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                 3           500  avgt   15   294136.644 ±  6709.506  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                 3          1000  avgt   15   599678.709 ±  5837.803  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10            10  avgt   15    17065.100 ±   241.715  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10           500  avgt   15   907448.421 ± 14203.990  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                10          1000  avgt   15  1906411.458 ± 26450.041  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20            10  avgt   15    33147.044 ±  1181.493  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20           500  avgt   15  1861946.674 ± 25145.868  ns/op
FetchResponseBenchmark.testSerializeFetchResponse                20          1000  avgt   15  4339003.253 ± 69576.240  ns/op
```

So consistently better, but not groundbreaking. That is fine since the main improvement is simplifying the generation of `Send` objects for all request types.