@hachikuji 2020-11-14
@chia7712 Thanks for reviews, merging to trunk. I will follow up with the suggestion about `AbstractResponse.parseResponse` as well as a couple other improvements.